+++ internal/controller/swiftimage/validate.go
@@ -17,6 +17,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 
 	imagev1alpha1 "github.com/kubeswift-io/kubeswift/api/image/v1alpha1"
+	"github.com/kubeswift-io/kubeswift/internal/controller/swiftguest"
 	"github.com/kubeswift-io/kubeswift/internal/names"
 )
 
@@ -57,9 +58,10 @@ func (r *SwiftImageReconciler) Validate(ctx context.Context, img *imagev1alpha1.
 						Spec: corev1.PodSpec{
 							RestartPolicy:                corev1.RestartPolicyNever,
 							AutomountServiceAccountToken: ptr.To(false),
+							ImagePullSecrets:             swiftguest.LauncherImagePullSecrets(),
 							Containers: []corev1.Container{{
 								Name:    "measure",
-								Image:   "ubuntu:22.04",
+								Image:   swiftguest.CloneJobImage(),
 								Command: []string{"sh", "-c", "cat /data/image.raw.size"},
 								VolumeMounts: []corev1.VolumeMount{{
 									Name:      "data",